# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: KMX EDC Console BFF API
  version: "1.0.0"
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    데이터스페이스 커넥터(EDC) 운영 콘솔의 BFF API.

    ## 인증 (Cookie 세션)
    - 로그인 성공 시 서버가 두 쿠키를 발급한다:
      - `kmx_token` — httpOnly JWT 세션(클라 JS 접근 불가), `SameSite=Strict`, prod 에서 `Secure`.
      - `kmx_csrf` — JS 가독 CSRF 이중제출 토큰.
    - 이후 모든 요청은 쿠키가 자동 첨부된다. 프로그램 클라이언트는
      `Authorization: Bearer <jwt>` 폴백도 사용할 수 있다.

    ## CSRF (double-submit)
    - 상태 변경 메서드(POST/PUT/PATCH/DELETE)는 `X-CSRF-Token` 헤더에 `kmx_csrf`
      쿠키 값을 되돌려 실어야 한다. 불일치 시 403.
    - 예외: `POST /api/auth/login` (자격증명으로 정당성이 검증되고 부작용이 세션 발급뿐).

    ## RBAC
    - 역할: `admin` > `operator` > `viewer`. 쓰기(생성/수정/삭제)는 operator 이상,
      플랫폼/설정은 admin. 권한 부족 시 403 `forbidden`.

    ## 에러 봉투
    - 모든 오류 응답은 `{ "error": string }` 를 기본으로 하며, EDC/DTR 프록시 오류는
      `source`("edc"|"dtr")와, 4xx 인 경우 `actionable: true` 를 덧붙일 수 있다.

    ## 멀티테넌시
    - 세션 JWT 의 `tenantId` 로 모든 리소스가 스코프된다. 커넥터 하위 경로는
      소유권 검사(requireConnectorOwnership)를 통과해야 한다(교차테넌트 접근 시 404).

    ## 페이지네이션
    - 목록 엔드포인트는 전체 배열을 반환하고 페이지네이션은 클라이언트에서 수행한다
      (usePagination + DataTablePagination). 서버측 page/limit 쿼리는 사용하지 않는다.
servers:
  - url: /
    description: 동일 오리진(콘솔이 BFF 와 같은 호스트에서 서빙됨)
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/connectors/{id}/transfers:
    parameters:
      - $ref: "#/components/parameters/ConnectorId"
    post:
      tags: [transfers]
      summary: 전송 프로세스 목록 조회(QuerySpec)
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QuerySpec" }
      responses:
        "200": { $ref: "#/components/responses/EdcObjectArray" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [transfers]
      summary: 전송 이력 일괄 정리(캐시)
      responses:
        "204": { description: 정리됨 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/connectors/{id}/transfers/{tpId}:
    parameters:
      - $ref: "#/components/parameters/ConnectorId"
      - $ref: "#/components/parameters/TransferId"
    get:
      tags: [transfers]
      summary: 전송 프로세스 단건 조회(상태)
      responses:
        "200": { $ref: "#/components/responses/EdcObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: kmx_token
      description: |
        httpOnly JWT 세션 쿠키. 로그인 시 발급된다. 상태 변경 요청은 추가로
        `X-CSRF-Token` 헤더에 `kmx_csrf` 쿠키 값을 실어야 한다(double-submit).
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 프로그램 클라이언트용 Authorization 헤더 폴백(쿠키 대체).
  parameters:
    ConnectorId:
      name: id
      in: path
      required: true
      description: 커넥터 ID(테넌트 소유권 검사 대상)
      schema: { type: string }
...
    TransferId:
      name: tpId
      in: path
      required: true
      description: 전송 프로세스 ID
      schema: { type: string }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 사람이 읽을 오류 메시지(민감정보는 서버에서 일반화됨)
        source:
          type: string
          enum: [edc, dtr]
          description: 프록시 상류 오류의 출처(있을 때만)
        actionable:
          type: boolean
          description: 4xx 상류 오류에서 상세를 그대로 노출해도 되는지(admin/operator 대상)
...
    QuerySpec:
      type: object
      description: |
        EDC Management API QuerySpec. 목록 필터·정렬·범위를 기술한다. 생략 시 서버
        기본값(대개 limit=50 이상)이 적용된다. 필드는 EDC 계약을 그대로 따른다.
      properties:
        offset: { type: integer }
        limit: { type: integer }
        sortField: { type: string }
        sortOrder: { type: string, enum: [ASC, DESC] }
        filterExpression:
          type: array
          items: { type: object, additionalProperties: true }
      additionalProperties: true
    EdcObject:
      type: object
      description: |
        EDC/DTR 상류가 반환하는 JSON-LD 오브젝트를 그대로 전달한 것. 스키마를 고정하지
        않는다(@context/@id/@type 등 JSON-LD 키 포함 가능). 정확한 형태는 EDC 0.17.0
        Management API 계약을 참조.
      additionalProperties: true
  responses:
    EdcObject:
      description: EDC/DTR 상류 오브젝트(pass-through)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/EdcObject" }
    EdcObjectArray:
      description: EDC/DTR 상류 오브젝트 배열(pass-through, 클라이언트 페이지네이션)
      content:
        application/json:
          schema:
            type: array
            items: { $ref: "#/components/schemas/EdcObject" }
...
    Unauthorized:
      description: 미인증(토큰 없음/무효/만료/무효화) 또는 세션 만료
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: 인증됐으나 권한 부족(RBAC) 또는 CSRF 실패
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: 리소스 없음 또는 교차테넌트 접근(소유권 검사 실패)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }