# server/routes/transfers.ts
// KMX EDC — Transfer Process Routes
//
// 크기/소요시간 추적 전략:
//   - transfer_metadata 테이블에 started_at, completed_at, size_bytes, user_completed 기록
//   - 전송 시작 시 started_at 기록
//   - 완료 처리 시 completed_at 기록 + user_completed=true → UI에서 COMPLETED로 오버레이
//   - POST /:id/transfers/:tpId/fetch → EDR로 실제 데이터 Pull, Content-Length → size_bytes 기록

import {
  Router,
  type Request,
  type Response,
  type NextFunction,
} from "express";
import { getConnector } from "../lib/connectorRegistry.js";
import {
  getEdcClient,
  withJsonLd,
  mapTransfer,
  type TransferMeta,
} from "../lib/edcClient.js";
import { getPool } from "../lib/db.js";
import { requireRole } from "../middleware/auth.js";
import { assertEndpointPublic } from "../middleware/validation.js";
import { pullEdrData, evictEdrTokens } from "../lib/edrRefresh.js";

const router = Router();
const writeGuard = requireRole("admin", "operator");

// EDR 데이터 pull 응답 크기 상한 — targetUrl 은 미신뢰 provider 데이터플레인이 반환한 EDR
// endpoint 라, 상한이 없으면 거대한 응답이 arraybuffer 로 전량 버퍼링돼 공유 BFF 를 OOM 시킬 수
// 있다(단일 프로세스 → 전 테넌트 동시 다운). env EDR_MAX_RESPONSE_BYTES 로 조정(기본 100MB).
const EDR_MAX_RESPONSE_BYTES = Number(
  process.env.EDR_MAX_RESPONSE_BYTES ?? 100 * 1024 * 1024
);

// 동기 JSON.parse/버퍼→문자열 변환은 큰 입력에서 이벤트루프를 블로킹(단일 프로세스 공유 BFF
// 프리즈)하므로, 이 상한 이하만 파싱한다. 초과 페이로드는 파싱을 생략하고 크기만 노출한다.
const EDR_MAX_PARSE_BYTES = Number(
  process.env.EDR_MAX_PARSE_BYTES ?? 8 * 1024 * 1024
);

/**
 * EDR endpoint(신뢰된 provider 데이터플레인 URL)에 선택적 하위 경로/쿼리를 안전하게 덧붙인다.
 * 절대 URL / protocol-relative(`//`) 는 호스트 변조 우려로 거부(null 반환).
 * 프록시 자산(DTR 등)을 하위 경로로 조회하기 위함.
 */
function appendProxyPath(
  endpoint: string,
  path?: unknown,
  query?: unknown
): string | null {
  let url = endpoint;
  if (typeof path === "string" && path.trim()) {
    const rel = path.trim();
    if (/^[a-z][a-z0-9+.-]*:\/\//i.test(rel) || rel.startsWith("//"))
      return null;
    url =
      endpoint.replace(/\/+$/, "") + (rel.startsWith("/") ? rel : `/${rel}`);
  }
  if (typeof query === "string" && query.trim()) {
    const q = query.trim().replace(/^[?&]+/, "");
    if (q) url += (url.includes("?") ? "&" : "?") + q;
  }
  return url;
}

async function resolveConnector(id: string) {
  const conn = await getConnector(id);
  if (!conn) throw new Error(`Connector ${id} not found`);
  return {
    conn,
    client: getEdcClient(conn.id, {
      managementUrl: conn.managementUrl,
      apiKey: conn.apiKey,
    }),
  };
}

/** 특정 커넥터의 메타데이터 Map 조회 */
async function getMetaMap(
  connectorId: string
): Promise<
  Map<string, TransferMeta & { hidden: boolean; last_state: string | null }>
> {
  const { rows } = await getPool().query(
    `SELECT transfer_id, user_completed, started_at, completed_at, size_bytes, fetch_duration_ms, hidden, last_state
     FROM transfer_metadata WHERE connector_id = $1`,
    [connectorId]
  );
  const map = new Map<
    string,
    TransferMeta & { hidden: boolean; last_state: string | null }
  >();
  for (const r of rows) {
    map.set(r.transfer_id, {
      user_completed: r.user_completed,
      started_at: r.started_at ?? null,
      completed_at: r.completed_at ?? null,
      size_bytes: r.size_bytes != null ? Number(r.size_bytes) : null,
      fetch_duration_ms:
        r.fetch_duration_ms != null ? Number(r.fetch_duration_ms) : null,
      hidden: r.hidden,
      last_state: r.last_state ?? null,
    });
  }
  return map;
}

// POST /:id/transfers — proxy to POST /v3/transferprocesses/request (list)
router.post(
  "/:id/transfers",
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);

      const [response, metaMap] = await Promise.all([
        client.post("/v3/transferprocesses/request", withJsonLd(req.body)),
        getMetaMap(connectorId),
      ]);

      const mapped = Array.isArray(response.data)
        ? response.data
            .map((raw: Record<string, unknown>) => {
              const id = (raw["@id"] as string) ?? "";
              return {
                ...mapTransfer(raw, metaMap.get(id)),
                _hidden: metaMap.get(id)?.hidden ?? false,
              };
            })
            .filter(t => !t._hidden)
            .map(({ _hidden: _, ...t }) => t)
        : response.data;

      // 성공률 KPI 용 최종 상태 기록 — 터미널(COMPLETED 1200/TERMINATED 1300/
      // DEPROVISIONED 1400) 도달 항목 중 미기록분만 1회 UPSERT. mapTransfer 의
      // 소비자완료 오버레이(name=COMPLETED)를 그대로 저장해 provider 측 오탐 실패를
      // 성공으로 집계한다. 협상 side-write(negotiations.ts)와 동일 패턴.
      if (Array.isArray(mapped)) {
        const toRecord = (
          mapped as Array<{ id: string; state: number; name: string }>
        ).filter(
          t => t.state >= 1200 && t.id && !metaMap.get(t.id)?.last_state
        );
        if (toRecord.length > 0) {
          await Promise.all(
            toRecord.map(t =>
              getPool().query(
                `INSERT INTO transfer_metadata (transfer_id, connector_id, completed_at, last_state)
                 VALUES ($1, $2, NOW(), $3)
                 ON CONFLICT (transfer_id, connector_id)
                 DO UPDATE SET
                   completed_at = COALESCE(transfer_metadata.completed_at, NOW()),
                   last_state = COALESCE(transfer_metadata.last_state, EXCLUDED.last_state)`,
                [t.id, connectorId, t.name]
              )
            )
          );
        }
      }

      res.json(mapped);
    } catch (error) {
      next(error);
    }
  }
);

// POST /:id/transfers/start — proxy to POST /v3/transferprocesses (start)
router.post(
  "/:id/transfers/start",
  writeGuard,
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);
      const { agreementId, counterPartyAddress, assetId, dataSink } =
        req.body ?? {};

      if (!agreementId) {
        res.status(400).json({ error: "agreementId is required" });
        return;
      }
      if (!counterPartyAddress) {
        res.status(400).json({
          error: "counterPartyAddress (provider DSP endpoint) is required",
        });
        return;
      }
      if (
        typeof counterPartyAddress !== "string" ||
        counterPartyAddress.length > 2048
      ) {
        res.status(400).json({
          error: "counterPartyAddress must be a string and within 2048 chars",
        });
        return;
      }
      const tpSsrfErr = await assertEndpointPublic(counterPartyAddress);
      if (tpSsrfErr) {
        res
          .status(400)
          .json({ error: `Rejected counterPartyAddress: ${tpSsrfErr}` });
        return;
      }

      // DSP 2025-1 endpoint 보정 (catalog/negotiations와 동일 패턴)
      let normalizedDspEndpoint = String(counterPartyAddress).replace(
        /\/+$/,
        ""
      );
      if (/\/api\/v1\/dsp$/.test(normalizedDspEndpoint)) {
        normalizedDspEndpoint = `${normalizedDspEndpoint}/2025-1`;
      }

      const sinkType: string = dataSink?.type ?? "HttpProxy";
      // sink 타입별 transferType/dataDestination 구성 — kmx-edc 지원 조합:
      //   HttpProxy → HttpData-PULL, HttpData → HttpData-PUSH, AmazonS3 → AmazonS3-PUSH.
      // (과거: S3 선택이 HttpData-PUSH 로 조용히 다운그레이드 — kmx-edc 0.17 정합.)
      let transferType: string;
      let dataDestination: Record<string, unknown>;
      if (sinkType === "AmazonS3") {
        const region = String(dataSink?.region ?? "").trim();
        const bucketName = String(dataSink?.bucketName ?? "").trim();
        if (!region || !bucketName) {
          res.status(400).json({
            error: "AmazonS3 sink requires region and bucketName",
          });
          return;
        }
        transferType = "AmazonS3-PUSH";
        dataDestination = {
          "@type": "DataAddress",
          type: "AmazonS3",
          region,
          bucketName,
          ...(dataSink?.objectName
            ? { objectName: String(dataSink.objectName) }
            : {}),
          ...(dataSink?.endpointOverride
            ? { endpointOverride: String(dataSink.endpointOverride) }
            : {}),
          ...(dataSink?.accessKeyId
            ? { accessKeyId: String(dataSink.accessKeyId) }
            : {}),
          ...(dataSink?.secretAccessKey
            ? { secretAccessKey: String(dataSink.secretAccessKey) }
            : {}),
        };
      } else if (sinkType === "HttpProxy") {
        transferType = "HttpData-PULL";
        dataDestination = { "@type": "DataAddress", type: "HttpProxy" };
      } else {
        transferType = "HttpData-PUSH";
        dataDestination = {
          "@type": "DataAddress",
          type: "HttpData",
          baseUrl: dataSink?.endpoint ?? "",
        };
      }

      const transferRequest: Record<string, unknown> = {
        "@context": { "@vocab": "https://w3id.org/edc/v0.0.1/ns/" },
        "@type": "TransferRequest",
        counterPartyAddress: normalizedDspEndpoint,
        contractId: agreementId,
        transferType,
        protocol: "dataspace-protocol-http:2025-1",
        dataDestination,
      };
      if (assetId) transferRequest["assetId"] = assetId;

      const response = await client.post(
        "/v3/transferprocesses",
        transferRequest
      );
      const tpId = (response.data as Record<string, unknown>)["@id"] as string;

      // 시작 시각 기록
      if (tpId) {
        await getPool().query(
          `INSERT INTO transfer_metadata (transfer_id, connector_id, started_at)
         VALUES ($1, $2, NOW())
         ON CONFLICT (transfer_id, connector_id) DO NOTHING`,
          [tpId, connectorId]
        );
      }

      res.json(response.data);
    } catch (error) {
      next(error);
    }
  }
);

// GET /:id/transfers/:tpId — proxy to GET /v3/transferprocesses/:tpId
router.get(
  "/:id/transfers/:tpId",
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const { client } = await resolveConnector(req.params.id);
      const response = await client.get(
        `/v3/transferprocesses/${req.params.tpId}`
      );
      res.json(response.data);
    } catch (error) {
      next(error);
    }
  }
);

// POST /:id/transfers/:tpId/fetch — EDR로 실제 데이터 Pull + 크기 기록
// 응답: { data, sizeBytes, contentType }
// write 가드: 실제 provider 데이터를 추출하므로 admin/operator만 허용(viewer 차단 — id 76).
router.post(
  "/:id/transfers/:tpId/fetch",
  writeGuard,
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const { tpId } = req.params;
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);

      // 1. EDR DataAddress 조회 (endpoint + authorization 토큰)
      const edrRes = await client.get(`/v3/edrs/${tpId}/dataaddress`);
      const edr = edrRes.data as Record<string, unknown>;
      const endpoint = edr["endpoint"] as string;
      const token = edr["authorization"] as string;

      if (!endpoint || !token) {
        res
          .status(404)
          .json({ error: "EDR not found or expired for this transfer" });
        return;
      }

      // 프록시 자산(예: DTR cx-taxo:DigitalTwinRegistry, proxyPath=true)은 EDR endpoint
      // 루트가 아니라 하위 경로로 조회해야 데이터가 나온다(루트 pull은 빈 응답 → size=—).
      // 선택적 path/query를 EDR endpoint에 안전하게 덧붙인다.
      const targetUrl = appendProxyPath(
        endpoint,
        req.body?.path,
        req.body?.query
      );
      if (targetUrl === null) {
        res
          .status(400)
          .json({ error: "path must be a relative sub-path (no scheme/host)" });
        return;
      }

      // SSRF 가드: EDR endpoint 는 미신뢰 provider 데이터플레인이 반환한 값이므로,
      // /start 의 counterPartyAddress 와 동일하게 사설/메타데이터 대역을 차단한다
      // (이 경로에 가드가 없으면 악성 provider 가 169.254.169.254 등으로 BFF 를 유도해
      //  내부망 도달 + EDR Bearer 토큰 유출 가능 — 데이터 Pull 경로 일관성 결손 보완).
      const fetchSsrfErr = await assertEndpointPublic(targetUrl);
      if (fetchSsrfErr) {
        res
          .status(502)
          .json({ error: `Rejected EDR data endpoint: ${fetchSsrfErr}` });
        return;
      }

      // 2. Provider Data Plane에서 실제 데이터 Pull — 소요시간 측정.
      // 액세스 토큰 만료(403) 시 관리 API /v3/edrs/{tpId}/refresh 로 자동 갱신 후 재시도.
      const fetchStart = Date.now();
      const dataRes = await pullEdrData(
        client,
        connectorId,
        tpId,
        targetUrl,
        edr,
        {
          responseType: "arraybuffer",
          timeout: 30_000,
          // SSRF: 리다이렉트 미추적 — provider 가 30x 로 내부/메타데이터 주소로 유도해 EDR
          // Bearer 토큰을 탈취하는 것 차단. + 응답 크기 상한으로 OOM(공유 BFF 다운) 방지.
          maxRedirects: 0,
          maxContentLength: EDR_MAX_RESPONSE_BYTES,
          maxBodyLength: EDR_MAX_RESPONSE_BYTES,
        }
      );
      const fetchDurationMs = Date.now() - fetchStart;

      const sizeBytes = dataRes.data?.length ?? 0;
      const contentType =
        (dataRes.headers["content-type"] as string) ??
        "application/octet-stream";

      // 3. size_bytes + fetch_duration_ms 기록 (UPSERT)
      await getPool().query(
        `INSERT INTO transfer_metadata (transfer_id, connector_id, size_bytes, fetch_duration_ms)
       VALUES ($1, $2, $3, $4)
       ON CONFLICT (transfer_id, connector_id)
       DO UPDATE SET size_bytes = EXCLUDED.size_bytes, fetch_duration_ms = EXCLUDED.fetch_duration_ms`,
        [tpId, connectorId, sizeBytes, fetchDurationMs]
      );

      // 4. 데이터 응답 (JSON이면 파싱, 아니면 raw). 단 대형 페이로드(>parse cap)는 동기
      //    JSON.parse/문자열 변환이 이벤트루프를 블로킹해 공유 BFF를 프리즈시키므로 파싱을
      //    생략하고 크기만 노출한다(previewOmitted).
      const raw = Buffer.from(dataRes.data);
      let body: unknown = null;
      let previewOmitted = false;
      if (raw.length <= EDR_MAX_PARSE_BYTES) {
        try {
          body = JSON.parse(raw.toString("utf-8"));
        } catch {
          body = raw.toString("utf-8");
        }
      } else {
        previewOmitted = true;
      }

      res.json({ data: body, sizeBytes, contentType, previewOmitted });
    } catch (error) {
      next(error);
    }
  }
);

// DELETE /:id/transfers — 전체 삭제 (STARTED는 terminate 후 hidden, 나머지는 바로 hidden)
router.delete(
  "/:id/transfers",
  writeGuard,
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);

      // 1. 현재 전송 목록 조회
      const response = await client.post(
        "/v3/transferprocesses/request",
        withJsonLd({})
      );
      const allTransfers: Array<{ id: string; state: string }> = Array.isArray(
        response.data
      )
        ? response.data.map((r: Record<string, unknown>) => ({
            id: (r["@id"] as string) ?? "",
            state: (r["state"] as string) ?? "",
          }))
        : [];

      // 2. STARTED 상태는 terminate 먼저 (종료되지 않으면 EDC 레코드가 남음)
      const started = allTransfers.filter(t => t.state === "STARTED");
      await Promise.allSettled(
        started.map(t =>
          client.post(`/v3/transferprocesses/${t.id}/terminate`, {
            "@context": { "@vocab": "https://w3id.org/edc/v0.0.1/ns/" },
            reason: "Deleted by user",
          })
        )
      );

      // 3. 모든 ID를 hidden=true로 표시 (UPSERT 배치)
      const ids = allTransfers.map(t => t.id).filter(Boolean);
      if (ids.length > 0) {
        await Promise.all(
          ids.map(id =>
            getPool().query(
              `INSERT INTO transfer_metadata (transfer_id, connector_id, hidden)
             VALUES ($1, $2, TRUE)
             ON CONFLICT (transfer_id, connector_id)
             DO UPDATE SET hidden = TRUE`,
              [id, connectorId]
            )
          )
        );
      }

      res.json({ deleted: ids.length });
    } catch (error) {
      next(error);
    }
  }
);

// POST /:id/transfers/:tpId/complete — 완료 처리
// 1. EDR로 데이터 Pull → size_bytes + fetch_duration_ms 기록 (아직 fetch 안 한 경우)
// 2. EDC /terminate 호출
// 3. completed_at + user_completed=true 기록 → 목록에서 COMPLETED 오버레이
router.post(
  "/:id/transfers/:tpId/complete",
  writeGuard,
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const { tpId } = req.params;
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);

      // 이미 fetch 기록이 있는지 확인
      const { rows } = await getPool().query(
        "SELECT fetch_duration_ms FROM transfer_metadata WHERE transfer_id = $1 AND connector_id = $2",
        [tpId, connectorId]
      );
      const alreadyFetched =
        rows.length > 0 && rows[0].fetch_duration_ms != null;

      // fetch 기록이 없으면 자동으로 Pull 수행
      if (!alreadyFetched) {
        try {
          const edrRes = await client.get(`/v3/edrs/${tpId}/dataaddress`);
          const edr = edrRes.data as Record<string, unknown>;
          const endpoint = edr["endpoint"] as string;
          const token = edr["authorization"] as string;

          if (endpoint && token) {
            // 프록시 자산(DTR 등)은 EDR 루트가 빈 응답 → /fetch와 동일하게 하위 경로 허용(id 77).
            const targetUrl = appendProxyPath(
              endpoint,
              req.body?.path,
              req.body?.query
            );
            // SSRF 가드: /fetch 와 동일하게 EDR endpoint 사설/메타데이터 대역 차단.
            // 무효 endpoint 면 데이터 Pull 만 건너뛰고 완료(terminate)는 계속 진행.
            if (
              targetUrl !== null &&
              (await assertEndpointPublic(targetUrl)) === null
            ) {
              const fetchStart = Date.now();
              // 만료(403) 시 관리 API refresh 로 자동 갱신 후 재시도(/fetch 와 동일).
              const dataRes = await pullEdrData(
                client,
                connectorId,
                tpId,
                targetUrl,
                edr,
                {
                  responseType: "arraybuffer",
                  timeout: 30_000,
                  // SSRF: 리다이렉트 미추적(EDR 토큰 유출 차단) + 크기 상한(OOM 방지).
                  maxRedirects: 0,
                  maxContentLength: EDR_MAX_RESPONSE_BYTES,
                  maxBodyLength: EDR_MAX_RESPONSE_BYTES,
                }
              );
              const fetchDurationMs = Date.now() - fetchStart;
              const rawSize = dataRes.data?.length ?? 0;
              // '0'(빈 응답)과 '미측정'을 구분 — 0이면 null로 남겨 통계 왜곡 방지.
              const sizeBytes = rawSize > 0 ? rawSize : null;

              await getPool().query(
                `INSERT INTO transfer_metadata (transfer_id, connector_id, size_bytes, fetch_duration_ms)
             VALUES ($1, $2, $3, $4)
             ON CONFLICT (transfer_id, connector_id)
             DO UPDATE SET size_bytes = EXCLUDED.size_bytes, fetch_duration_ms = EXCLUDED.fetch_duration_ms`,
                [tpId, connectorId, sizeBytes, fetchDurationMs]
              );
            }
          }
        } catch {
          // EDR 만료 등으로 fetch 실패해도 완료 처리는 계속 진행
        }
      }

      // EDC terminate
      await client.post(`/v3/transferprocesses/${tpId}/terminate`, {
        "@context": { "@vocab": "https://w3id.org/edc/v0.0.1/ns/" },
        reason: "Completed by consumer",
      });
      // 종료된 전송의 공유 저장소 토큰을 즉시 정리(불필요한 자격증명 상주 최소화).
      // 정리 실패가 완료 처리를 막지 않도록 fire-and-forget(보존정리 prune 가 최종 백스톱).
      void evictEdrTokens(connectorId, tpId).catch(() => {});

      // 완료 메타 기록
      await getPool().query(
        `INSERT INTO transfer_metadata (transfer_id, connector_id, completed_at, user_completed)
       VALUES ($1, $2, NOW(), TRUE)
       ON CONFLICT (transfer_id, connector_id)
       DO UPDATE SET completed_at = NOW(), user_completed = TRUE`,
        [tpId, connectorId]
      );

      res.status(204).end();
    } catch (error) {
      next(error);
    }
  }
);

// POST /:id/transfers/:tpId/terminate — 강제 종료
router.post(
  "/:id/transfers/:tpId/terminate",
  writeGuard,
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const { tpId } = req.params;
      const connectorId = req.params.id;
      const { client } = await resolveConnector(connectorId);
      const { reason } = req.body ?? {};

      await client.post(`/v3/transferprocesses/${tpId}/terminate`, {
        "@context": { "@vocab": "https://w3id.org/edc/v0.0.1/ns/" },
        reason: reason ?? "Terminated by user",
      });

      await getPool().query(
        `INSERT INTO transfer_metadata (transfer_id, connector_id, completed_at)
       VALUES ($1, $2, NOW())
       ON CONFLICT (transfer_id, connector_id)
       DO UPDATE SET completed_at = NOW()`,
        [tpId, connectorId]
      );

      res.status(204).end();
    } catch (error) {
      next(error);
    }
  }
);

export default router;